refactoring the landing header out of Landing.js
